requirements.txt
arrow==1.2.3
asgiref==3.6.0
blessed==1.19.1
Django==4.1.5
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
wcwidth==0.2.6
